+++ src/main/java/org/bytedeco/javacpp/tools/InfoMap.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2017 Samuel Audet
+ * Copyright (C) 2014-2018 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -93,6 +93,7 @@ public class InfoMap extends HashMap<String,List<Info>> {
         .put(new Info("boost::shared_ptr", "std::shared_ptr").annotations("@SharedPtr"))
         .put(new Info("boost::movelib::unique_ptr", "std::unique_ptr").annotations("@UniquePtr"))
         .put(new Info("std::string").annotations("@StdString").valueTypes("BytePointer", "String"))
+        .put(new Info("std::wstring").annotations("@StdWString").valueTypes("CharPointer", "IntPointer"))
         .put(new Info("std::vector").annotations("@StdVector"))
 
         .put(new Info("abstract").javaNames("_abstract"))