+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Enhance `InfoMap` and `StringAdapter` with default mappings and casts for `std::wstring`
  * Templatize `StringAdapter` to allow other character types like `wchar_t` and add corresponding `@StdWString` annotation
  * Prevent `Loader` from creating symbolic links to rename libraries, which does not always work
  * Fix memory leak that occurs with "org.bytedeco.javacpp.nopointergc" ([issue #239](https://github.com/bytedeco/javacpp/issues/239))